+++ CHANGELOG.md
@@ -1,10 +1,20 @@
-<a name="0.1.0"></a>
-# 0.1.0 (2018-01-28)
+<a name="0.2.0"></a>
+# [0.2.0](https://github.com/Beg-in/3d/compare/0.1.0...0.2.0) (2018-01-29)
 
 
 ### Features
 
-* **plugin:** create Vue and Vuex plugins ([2ef766b](https://github.com/Beg-in/3d/commit/2ef766b))
+* **entity:** add entity and some basic meshes ([ebb4a76](https://github.com/Beg-in/3d/commit/ebb4a76))
+
+
+
+<a name="0.1.0"></a>
+# [0.1.0](https://github.com/Beg-in/3d/compare/1c8615b...0.1.0) (2018-01-28)
+
+
+### Features
+
+* **plugin:** create Vue and Vuex plugins ([1c8615b](https://github.com/Beg-in/3d/commit/1c8615b))